+++ EnsNetcode/Netcode/Common/EnsInstance.cs
@@ -29,8 +29,8 @@ public class EnsInstance
     public static Action OnConnectionRejected;
     public static Action OnServerConnect;
     public static Action OnServerDisconnect;
-    public static Action<int> OnClientEnter;//有新用户连接到服务器时触发(新用户自身不调用)
-    public static Action<int> OnClientExit;//有用户与服务器断开时调用(断开的用户自身不调用)
+    public static Action<short> OnClientEnter;//有新用户连接到服务器时触发(新用户自身不调用)
+    public static Action<short> OnClientExit;//有用户与服务器断开时调用(断开的用户自身不调用)
     public static Action OnAuthorityChanged;
 
     public static Action OnCreateRoom { get => CreateRoom.OnCreateRoom; set=>CreateRoom.OnCreateRoom=value; }
